Add exponents
This is where associativity comes into play because the ^ is opposite of the others.
This also brought out differences in the expected results to match the kata example and
the example on the Shunting-Yard algorithm wiki page.

I think all the cases I had before were equivalent, just different. For example,
    acd++
    ac+d+
result in the same output:
    123++ -> (2+3)+1 -> 6
    12+3+ -> (1+2)+3 -> 6

Also the same:
    cdbae-/*+ -> 34215-/*+ -> 3+(4*(2/(1-5))) -> 5
    cdb*ae-/+ -> 342*15-/+ -> 3+(4*2)/(1-5) -> 5

However, once the exponent enters the game with the opposite associativity, these minor differences
become significant and can cause the exponent to be applied to the wrong value.

That's why the previous test cases had to be changed.